requirements.txt
# AI-Powered Phishing Detection & Generation System
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
joblib>=1.3.0
PyYAML>=6.0
jupyter>=1.0.0
ipykernel>=6.25.0
flask>=3.0.0
